requirements.txt
aiohttp>=3.5.4
aioify>=0.3.1
ddp_asyncio>=0.3.0
dmsclient>=1.5.1
petname>=2.2
raven>=6.10.0
rocketbot==0.2.7
rocketchat_API>=0.6.31
tzlocal>=1.5.1
ftfbroker==0.0.3
atlassian-python-api